Imports: move adding command line scopes to main
