+++ lib/spack/spack/environment/__init__.py
@@ -471,6 +471,7 @@ from .environment import (
     activate,
     active,
     active_environment,
+    add_path_scopes,
     all_environment_names,
     all_environments,
     as_env_dir,
@@ -507,6 +508,7 @@ __all__ = [
     "activate",
     "active",
     "active_environment",
+    "add_path_scopes",
     "all_environment_names",
     "all_environments",
     "as_env_dir",
